Track Token Generation Time
This prints the token generation time after each run.
There is a possibility that the sampling time and prompt evaluation
time are getting added into the mix. I think there is a bug in timing
calculation based on the comments written.

In any case, the sampling time and prompt eval time are so low, that
they could be ignored for most cases and do not deviate dramtically from
the overall results.

Also, llama.cpp prints its own version of token generation time called
"eval time" (which I think is buggy) if in case a fallback value is needed.